requirements.txt
crewai==1.9.2
langchain-openai==0.3.23
python-dotenv==1.1.1
